+++ storybook/stories/staking/StakePools.stories.js
@@ -93,5 +93,8 @@ export const StakePoolsStory = (props: Props) => (
     onSmashSettingsClick={action('onSmashSettingsClick')}
     smashServerUrl="https://smash.cardano-mainnet.iohk.io"
     maxDelegationFunds={maxDelegationFunds}
+    filterOptions={{}}
+    populatedFilterOptions={{}}
+    onFilter={() => null}
   />
 );